active-projects/db_historical_updates/requirements.txt
numpy==1.14.6
pandas==1.2.4
pyjarowinkler==1.8
python_dateutil==2.8.1
scipy==1.6.3
SQLAlchemy==1.4.15
